Fix: changes done after discussion
