+++ consent-server/internal/consent/expiration_job.go
@@ -29,7 +29,7 @@ import (
 // ExpirationService is the minimal interface required by the expiration job.
 // Both *consentService and test doubles satisfy this interface.
 type ExpirationService interface {
-	GetExpiredConsents(ctx context.Context, nowMs int64, expirableStatuses []string) ([]model.Consent, *serviceerror.ServiceError)
+	GetExpiredConsents(ctx context.Context, currentTimeMs int64, expirableStatuses []string) ([]model.Consent, *serviceerror.ServiceError)
 	ExpireConsent(ctx context.Context, consent *model.Consent, orgID string) error
 }
